# apps/desktop-companion/openapi.yaml
openapi: 3.1.0
info:
  title: FluxBoard Livestream Copilot API
  description: |
    REST API for the FluxBoard Livestream Copilot desktop companion service.

    FluxBoard provides real-time livestream capture, AI-powered content generation,
    OBS Studio integration, clip management, social media export, and video streaming
    for content creators.

    ## Authentication

    Most endpoints require authentication via one of two methods:

    1. **Bearer JWT** - Pass an access token in the `Authorization` header.
    2. **API Key** - Pass an API key in the `x-api-key` header.

    Public endpoints (health checks, OBS status, event reads) do not require authentication
    during development.

    ## WebSocket Events

    The server exposes a WebSocket on a configurable port (default 9001). Clients receive
    real-time `EventEnvelope` messages for session lifecycle, transcripts, clips, outputs,
    and trigger detections. Connect and listen for JSON messages.

    Event types: `SESSION_START`, `SESSION_END`, `TRANSCRIPT_SEGMENT`, `MOMENT_MARKER`,
    `CLIP_INTENT_START`, `CLIP_INTENT_END`, `ARTIFACT_CLIP_CREATED`, `ARTIFACT_FRAME_CREATED`,
    `OUTPUT_CREATED`, `OUTPUT_VALIDATED`, `AUTO_TRIGGER_DETECTED`, `CLIP_QUEUE_UPDATED`.

    ## Rate Limiting

    Rate-limited endpoints return standard `RateLimit-*` headers. When exceeded, the server
    responds with `429 Too Many Requests`.

    | Endpoint Group        | Limit              |
...

  version: 1.0.0
  contact:
...
  license:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/v1/auth/login or /api/v1/auth/refresh
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Static API key for service-to-service communication
  schemas:
...
    Clip:
      type: object
      properties:
        id:
          type: string
        artifactId:
          type: string
        sessionId:
          type: string
        path:
          type: string
        t0:
          type: number
          description: Start time in seconds from session start
        t1:
          type: number
          description: End time in seconds from session start
        duration:
          type: number
          description: Duration in seconds (t1 - t0)
        thumbnailId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ── ClipQueueItem ──────────────────────────────────────────────────────
...
    SocialPlatform:
      type: string
      enum: [YOUTUBE, TWITTER, TIKTOK, LINKEDIN, INSTAGRAM, FACEBOOK, THREADS, BLUESKY]
  parameters:
...
    SessionIdParam:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Session ID
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 100
        minimum: 1
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0

# =============================================================================
# PATHS
# =============================================================================
paths:
...
  /api/clips/{id}:
    get:
      tags: [Clips]
      summary: Get clip by ID or artifact ID
      operationId: getClip
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Clip database ID or artifact ID
      responses:
        "200":
          description: Clip details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      clip:
                        $ref: '#/components/schemas/Clip'
        "404":
          description: Clip not found
    delete:
      tags: [Clips]
      summary: Delete a clip
      operationId: deleteClip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Clip deleted
        "404":
          description: Clip not found
  /api/sessions/{sessionId}/clips:
    get:
      tags: [Sessions, Clips]
      summary: List clips for a session
      operationId: listSessionClips
      security: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        "200":
          description: Paginated clip list
        "404":
          description: Session not found

  # ── Clip Queue ───────────────────────────────────────────────────────────
...
  /api/clip-queue/{id}:
    get:
      tags: [ClipQueue]
      summary: Get clip queue item
      operationId: getClipQueueItem
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Queue item details
        "404":
          description: Item not found
    patch:
      tags: [ClipQueue]
      summary: Update clip queue item title
      operationId: updateClipQueueItem
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
      responses:
        "200":
          description: Item updated
    delete:
      tags: [ClipQueue]
      summary: Delete/cancel a clip queue item
      operationId: deleteClipQueueItem
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item deleted
        "400":
          description: Cannot delete item currently processing
  /api/clip-queue/{id}/retry:
    post:
      tags: [ClipQueue]
      summary: Retry a failed clip queue item
      operationId: retryClipQueueItem
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item reset to pending for reprocessing
        "400":
          description: Can only retry failed items
  /api/sessions/{sessionId}/clip-queue:
    get:
      tags: [Sessions, ClipQueue]
      summary: List clip queue items for a session
      operationId: listSessionClipQueue
      security: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, RECORDING, PROCESSING, COMPLETED, FAILED]
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        "200":
          description: Queue items with stats and pagination
  /api/sessions/{sessionId}/clip-queue/manual:
    post:
      tags: [Sessions, ClipQueue]
      summary: Create a manual clip from last N seconds
      operationId: createManualClip
      security: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                durationSeconds:
                  type: integer
                  minimum: 5
                  maximum: 300
                  default: 30
                title:
                  type: string
                  maxLength: 200
      responses:
        "201":
          description: Manual clip queued
        "404":
          description: Session not found

  # ── OBS ──────────────────────────────────────────────────────────────────
...
  /api/v1/export/clip:
    post:
      tags: [Export]
      summary: Export a video clip with format conversion
      operationId: exportClip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clipId, format]
              properties:
                clipId:
                  type: string
                format:
                  type: string
                  enum: [MP4, WEBM, GIF, MOV]
                platform:
                  $ref: '#/components/schemas/SocialPlatform'
                options:
                  type: object
                  properties:
                    quality:
                      type: string
                      enum: [low, medium, high, original]
                    generateThumbnail:
                      type: boolean
                    addWatermark:
                      type: boolean
                    optimizeForPlatform:
                      type: boolean
                    targetAspectRatio:
                      type: string
      responses:
        "201":
          description: Clip export started
...
  /clip:
    post:
      tags: [Legacy]
      summary: Create a clip from replay buffer
      operationId: createClipLegacy
      security: []
      responses:
        "200":
          description: Clip created with artifact ID and paths
        "404":
          description: No active session
...